Add feature importance and utils for reading df
